+++ src/components/scenes/GuiPluginListScene.tsx
@@ -6,7 +6,7 @@ import FastImage from 'react-native-fast-image'
 import Animated from 'react-native-reanimated'
 import { sprintf } from 'sprintf-js'
 
-import { showBackupForTransferModal } from '../../actions/BackupModalActions'
+import { checkAndShowLightBackupModal } from '../../actions/BackupModalActions'
 import { getDeviceSettings, writeDeveloperPluginUri } from '../../actions/DeviceSettingsActions'
 import { NestedDisableMap } from '../../actions/ExchangeInfoActions'
 import { readSyncedSettings, SyncedAccountSettings, updateOneSetting, writeSyncedSettings } from '../../actions/SettingsActions'
@@ -229,11 +229,9 @@ class GuiPluginList extends React.PureComponent<Props, State> {
     const { pluginId, paymentType, deepQuery = {} } = listRow
     const plugin = guiPlugins[pluginId]
 
-    // Don't allow light accounts to enter plugins
-    if (account.username == null) {
-      showBackupForTransferModal(() => navigation.navigate('upgradeUsername', {}))
-      return
-    }
+    // Don't allow light accounts to enter sell plugins
+    const direction = this.getSceneDirection()
+    if (direction === 'buy' && checkAndShowLightBackupModal(account, navigation)) return
 
     // Grab a custom URI if necessary:
     let { deepPath = undefined } = listRow
@@ -260,8 +258,6 @@ class GuiPluginList extends React.PureComponent<Props, State> {
         writeDeveloperPluginUri(deepPath).catch(error => showError(error))
       }
     }
-
-    const direction = this.getSceneDirection()
     if (plugin.nativePlugin != null) {
       const disableProviders = disablePlugins[pluginId]